+++ src/billing/Payment.h
@@ -7,7 +7,7 @@ public:
     Payment(int tableId, float amount);
     float getAmountOfPayment() const;
     void setAmountOfPayment(float amount);
-    float gettableId() const;
+    int gettableId() const;
     void settableId(int tableId);
     virtual void paymentDetails() const;
 private:
